+++ src/main/kotlin/app/revanced/patches/photomath/misc/unlock/bookpoint/EnableBookpointPatch.kt
@@ -1,11 +1,11 @@
-package app.revanced.patches.photomath.misc.bookpoint
+package app.revanced.patches.photomath.misc.unlock.bookpoint
 
 import app.revanced.util.exception
 import app.revanced.patcher.data.BytecodeContext
 import app.revanced.patcher.extensions.InstructionExtensions.replaceInstructions
 import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.annotation.Patch
-import app.revanced.patches.photomath.misc.bookpoint.fingerprints.IsBookpointEnabledFingerprint
+import app.revanced.patches.photomath.misc.unlock.bookpoint.fingerprints.IsBookpointEnabledFingerprint
 
 @Patch(description = "Enables textbook access")
 internal object EnableBookpointPatch : BytecodePatch(
